fold timer into utilities, clean up ruff warnings
